Create buch_images_cropper.py
The function that crops all the images in the selected directory and with the selected format (either png or jpg, other formats were not checked), according to the entered coordinates in pixels (left, top, right, bottom). The coordinates can be found in MS Paint while opening the image with it. The solution should work well while only one image needs to be cropped. Basically, it crops all the images in the selected directory. Note that it assumes that all the images in the selected directory have the same size in pixels. Possible causes of an error are:
1) Wrong directory path.
2) Wrong image file type.
3) Images are not of the same size.